Convert opcodes to Enums
