+++ setup.cfg
@@ -16,6 +16,7 @@ packages =
     ethereum
     ethereum/frontier
     ethereum/frontier/vm
+    ethereum/frontier/vm/instructions
 
 package_dir =
     =src
