+++ src/ethereum/frontier/vm/instructions/stack.py
@@ -1,0 +1,175 @@
+"""
+Ethereum Virtual Machine (EVM) stack Instructions
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. contents:: Table of Contents
+    :backlinks: none
+    :local:
+
+Introduction
+------------
+
+Implementations of the EVM stack related instructions.
+"""
+
+from functools import partial
+
+from ethereum.base_types import U256
+
+from .. import Evm
+from ..gas import GAS_VERY_LOW, subtract_gas
+from ..stack import push
+
+
+def push_n(evm: Evm, num_bytes: int) -> None:
+    """
+    Pushes a N-byte immediate onto the stack.
+
+    Parameters
+    ----------
+    evm :
+        The current EVM frame.
+
+    num_bytes :
+        The number of immediate bytes to be read from the code and pushed to
+        the stack.
+
+    Raises
+    ------
+    StackOverflowError
+        If `len(stack)` is equals `1024`.
+    OutOfGasError
+        If `evm.gas_left` is less than `GAS_VERY_LOW`.
+    """
+    assert evm.pc + num_bytes < len(evm.code)
+    evm.gas_left = subtract_gas(evm.gas_left, GAS_VERY_LOW)
+
+    data_to_push = U256.from_be_bytes(
+        evm.code[evm.pc + 1 : evm.pc + num_bytes + 1]
+    )
+    push(evm.stack, data_to_push)
+
+    evm.pc += num_bytes
+
+
+def dup_n(evm: Evm, item_number: int) -> None:
+    """
+    Duplicate the Nth stack item (from top of the stack) to the top of stack.
+
+    Parameters
+    ----------
+    evm :
+        The current EVM frame.
+
+    item_number :
+        The stack item number (0-indexed from top of stack) to be duplicated
+        to the top of stack.
+
+    Raises
+    ------
+    OutOfGasError
+        If `evm.gas_left` is less than `GAS_VERY_LOW`.
+    """
+    assert item_number < len(evm.stack)
+    evm.gas_left = subtract_gas(evm.gas_left, GAS_VERY_LOW)
+
+    data_to_duplicate = evm.stack[len(evm.stack) - 1 - item_number]
+    push(evm.stack, data_to_duplicate)
+
+
+def swap_n(evm: Evm, item_number: int) -> None:
+    """
+    Swap the 1st and Nth items in the stack. All items are 0-indexed from the
+    top of the stack.
+
+    Parameters
+    ----------
+    evm :
+        The current EVM frame.
+
+    item_number :
+        The stack item number (0-indexed from top of stack) to be swapped
+        with the top of stack element.
+
+    Raises
+    ------
+    OutOfGasError
+        If `evm.gas_left` is less than `GAS_VERY_LOW`.
+    """
+    assert item_number < len(evm.stack)
+    evm.gas_left = subtract_gas(evm.gas_left, GAS_VERY_LOW)
+
+    top_element_idx = len(evm.stack) - 1
+    nth_element_idx = len(evm.stack) - 1 - item_number
+    evm.stack[top_element_idx], evm.stack[nth_element_idx] = (
+        evm.stack[nth_element_idx],
+        evm.stack[top_element_idx],
+    )
+
+
+push1 = partial(push_n, num_bytes=1)
+push2 = partial(push_n, num_bytes=2)
+push3 = partial(push_n, num_bytes=3)
+push4 = partial(push_n, num_bytes=4)
+push5 = partial(push_n, num_bytes=5)
+push6 = partial(push_n, num_bytes=6)
+push7 = partial(push_n, num_bytes=7)
+push8 = partial(push_n, num_bytes=8)
+push9 = partial(push_n, num_bytes=9)
+push10 = partial(push_n, num_bytes=10)
+push11 = partial(push_n, num_bytes=11)
+push12 = partial(push_n, num_bytes=12)
+push13 = partial(push_n, num_bytes=13)
+push14 = partial(push_n, num_bytes=14)
+push15 = partial(push_n, num_bytes=15)
+push16 = partial(push_n, num_bytes=16)
+push17 = partial(push_n, num_bytes=17)
+push18 = partial(push_n, num_bytes=18)
+push19 = partial(push_n, num_bytes=19)
+push20 = partial(push_n, num_bytes=20)
+push21 = partial(push_n, num_bytes=21)
+push22 = partial(push_n, num_bytes=22)
+push23 = partial(push_n, num_bytes=23)
+push24 = partial(push_n, num_bytes=24)
+push25 = partial(push_n, num_bytes=25)
+push26 = partial(push_n, num_bytes=26)
+push27 = partial(push_n, num_bytes=27)
+push28 = partial(push_n, num_bytes=28)
+push29 = partial(push_n, num_bytes=29)
+push30 = partial(push_n, num_bytes=30)
+push31 = partial(push_n, num_bytes=31)
+push32 = partial(push_n, num_bytes=32)
+
+dup1 = partial(dup_n, item_number=0)
+dup2 = partial(dup_n, item_number=1)
+dup3 = partial(dup_n, item_number=2)
+dup4 = partial(dup_n, item_number=3)
+dup5 = partial(dup_n, item_number=4)
+dup6 = partial(dup_n, item_number=5)
+dup7 = partial(dup_n, item_number=6)
+dup8 = partial(dup_n, item_number=7)
+dup9 = partial(dup_n, item_number=8)
+dup10 = partial(dup_n, item_number=9)
+dup11 = partial(dup_n, item_number=10)
+dup12 = partial(dup_n, item_number=11)
+dup13 = partial(dup_n, item_number=12)
+dup14 = partial(dup_n, item_number=13)
+dup15 = partial(dup_n, item_number=14)
+dup16 = partial(dup_n, item_number=15)
+
+swap1 = partial(swap_n, item_number=1)
+swap2 = partial(swap_n, item_number=2)
+swap3 = partial(swap_n, item_number=3)
+swap4 = partial(swap_n, item_number=4)
+swap5 = partial(swap_n, item_number=5)
+swap6 = partial(swap_n, item_number=6)
+swap7 = partial(swap_n, item_number=7)
+swap8 = partial(swap_n, item_number=8)
+swap9 = partial(swap_n, item_number=9)
+swap10 = partial(swap_n, item_number=10)
+swap11 = partial(swap_n, item_number=11)
+swap12 = partial(swap_n, item_number=12)
+swap13 = partial(swap_n, item_number=13)
+swap14 = partial(swap_n, item_number=14)
+swap15 = partial(swap_n, item_number=15)
+swap16 = partial(swap_n, item_number=16)
